requirements.txt
Django==2.0.7
icalendar==4.0.2
pytz==2018.5
requests==2.19.1
djangorestframework==3.8.2
